Add some CSS style and forecast class
