+++ package.json
@@ -7,6 +7,7 @@
     "build": "node_modules/.bin/webpack --mode production",
     "serve": "webpack-dev-server --mode development"
   },
+  
   "author": "",
   "license": "ISC",
   "devDependencies": {
